+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 1. Remove config.async, the reason is because [THRIFT-2376: nodejs: allow Promise style calls for client and server](https://issues.apache.org/jira/browse/THRIFT-2376).
 2. Support ping and can config ping method in config.js.
+3. Support get client by thriftilyManger.get('alias name').
 
 ## v0.1.1 2019.5.8 GMT+0800